refactor: Extract AppearanceControl
