+++ fireplace-app/src/main/java/io/github/bric3/fireplace/FirePlaceMain.java
@@ -9,18 +9,8 @@
  */
 package io.github.bric3.fireplace;
 
-import com.formdev.flatlaf.FlatDarculaLaf;
-import com.formdev.flatlaf.FlatIntelliJLaf;
-import com.formdev.flatlaf.FlatLaf;
-import com.formdev.flatlaf.extras.FlatAnimatedLafChange;
 import com.formdev.flatlaf.util.SystemInfo;
-import com.github.weisj.darklaf.platform.decorations.DecorationsColorProvider;
-import com.github.weisj.darklaf.platform.decorations.ExternalLafDecorator;
-import com.github.weisj.darklaf.platform.preferences.SystemPreferencesManager;
-import io.github.bric3.fireplace.core.ui.Colors;
 import io.github.bric3.fireplace.core.ui.JScrollPaneWithButton;
-import io.github.bric3.fireplace.icons.darkMode_moon;
-import io.github.bric3.fireplace.icons.darkMode_sun;
 import io.github.bric3.fireplace.ui.FrameResizeLabel;
 import io.github.bric3.fireplace.ui.HudPanel;
 import io.github.bric3.fireplace.ui.debug.AssertiveRepaintManager;
@@ -71,7 +61,6 @@ public class FirePlaceMain {
         if (Boolean.getBoolean("fireplace.swing.debug") || Boolean.getBoolean("fireplace.debug")) {
             System.getProperties().forEach((k, v) -> System.out.println(k + " = " + v));
         }
-        setupLaF();
 
         if (Boolean.getBoolean("fireplace.swing.debug")) {
             if (Objects.equals(System.getProperty("fireplace.swing.debug.thread.violation.checker"), "IJ")) {
@@ -82,8 +71,6 @@ public class FirePlaceMain {
             EventDispatchThreadHangMonitor.initMonitoring();
         }
         SwingUtilities.invokeLater(() -> {
-
-
             var openedFileLabel = new JTextField("");
             {
                 openedFileLabel.setBorder(BorderFactory.createEmptyBorder(5, 5, 5, 5));
@@ -130,7 +117,7 @@ public class FirePlaceMain {
                 openedFileLabel.setBorder(BorderFactory.createEmptyBorder(30, 5, 5, 5));
             }
             topPanel.add(openedFileLabel, BorderLayout.CENTER);
-            topPanel.add(AppearanceControl.INSTANCE.getComponent(), BorderLayout.EAST);
+            topPanel.add(AppearanceControl.getComponent(), BorderLayout.EAST);
 
             var mainPanel = new JPanel(new BorderLayout());
             mainPanel.add(topPanel, BorderLayout.NORTH);
@@ -144,22 +131,15 @@ public class FirePlaceMain {
             appLayers.setLayout(new OverlayLayout(appLayers));
             appLayers.setOpaque(false);
             appLayers.setVisible(true);
-            appLayers.add(mainPanel, JLayeredPane.DEFAULT_LAYER);
-            appLayers.add(hudPanel.getComponent(), JLayeredPane.MODAL_LAYER);
+            appLayers.add(mainPanel, JLayeredPane.PALETTE_LAYER);
+            // appLayers.add(hudPanel.getComponent(), JLayeredPane.MODAL_LAYER);
             appLayers.add(frameResizeLabel.getComponent(), JLayeredPane.POPUP_LAYER);
 
             JfrFilesDropHandler.install(jfrBinder::load, appLayers, hudPanel.getDnDTarget());
 
             var frame = new JFrame("FirePlace");
-            final var rootPane = frame.getRootPane();
-            if (SystemInfo.isMacOS) {
-                // allows to place swing components on the whole window
-                rootPane.putClientProperty("apple.awt.fullWindowContent", true);
-                // makes the title bar transparent
-                rootPane.putClientProperty("apple.awt.transparentTitleBar", true);
-            }
 
-            installAppIcon(frame);
+
             frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
             frame.setSize(new Dimension(1000, 600));
             frame.getContentPane().add(appLayers);
@@ -176,170 +156,13 @@ public class FirePlaceMain {
             });
 
             frame.getGraphicsConfiguration(); // get active screen
+
+            AppearanceControl.installAppIcon(frame);
+            AppearanceControl.install(frame);
+
             frame.setVisible(true);
         });
     }
 
-    private static void setupLaF() {
-        if (SystemInfo.isLinux) {
-            // most linux distros have ugly font rendering, but these here can fix that:
-            System.setProperty("awt.useSystemAAFontSettings", "on");
-            System.setProperty("swing.aatext", "true");
-            System.setProperty("sun.java2d.xrender", "true");
-        }
 
-        if (SystemInfo.isMacOS) {
-            System.setProperty("apple.laf.useScreenMenuBar", "true"); // moves menu bar from JFrame window to top of screen
-            System.setProperty("apple.awt.application.name", "FirePlace"); // application name used in screen menu bar
-            // appearance of window title bars
-            // possible values:
-            //   - "system": use current macOS appearance (light or dark)
-            //   - "NSAppearanceNameAqua": use light appearance
-            //   - "NSAppearanceNameDarkAqua": use dark appearance
-            //            System.setProperty("apple.awt.application.appearance", "NSAppearanceNameDarkAqua");
-            System.setProperty("apple.awt.application.appearance", "system");
-        }
-
-        AppearanceControl.INSTANCE.install();
-    }
-
-    private static class AppearanceControl {
-        public static final AppearanceControl INSTANCE = new AppearanceControl();
-        public static final String TO_APPEARANCE = "CURRENT_MODE";
-        public static final String TO_LIGHT_LAF = "LIGHT";
-        public static final String TO_DARK_LAF = "DARK";
-        public final Runnable SYNC_THEME_CHANGER;
-        private final SystemPreferencesManager manager;
-
-        public AppearanceControl() {
-            manager = new SystemPreferencesManager();
-
-            SYNC_THEME_CHANGER = () -> {
-                // System.out.println(">>>> theme preference changed = " + ThemePreferencesHandler.getSharedInstance().getPreferredThemeStyle());
-                switch (manager.getPreferredThemeStyle().getColorToneRule()) {
-                    case DARK:
-                        Colors.setDarkMode(true);
-                        FlatDarculaLaf.setup();
-                        break;
-                    case LIGHT:
-                        Colors.setDarkMode(false);
-                        FlatIntelliJLaf.setup();
-                        break;
-                }
-
-                FlatLaf.updateUI();
-                FlatAnimatedLafChange.hideSnapshotWithAnimation();
-            };
-            manager.addListener(style -> SYNC_THEME_CHANGER.run());
-        }
-
-        void install() {
-            ExternalLafDecorator.instance().install();
-
-            ExternalLafDecorator.instance().setColorProvider(new DecorationsColorProvider() {
-                @Override
-                public Color backgroundColor() {
-                    return UIManager.getColor("TitlePane.background");
-                }
-
-                @Override
-                public Color activeForegroundColor() {
-                    // ignored on macOs
-                    return UIManager.getColor("TitlePane.foreground");
-                }
-
-                @Override
-                public Color inactiveForegroundColor() {
-                    return Color.ORANGE;
-                }
-
-                @Override
-                public TitleColor windowTitleColor() {
-                    return FlatLaf.isLafDark() ? TitleColor.LIGHT : TitleColor.DARK;
-                }
-            });
-            SYNC_THEME_CHANGER.run();
-            toggleSync(true);
-        }
-
-        void toggleSync(boolean sync) {
-            try {
-                manager.enableReporting(sync);
-            } catch (Exception e) {
-                System.getLogger(AppearanceControl.class.getName()).log(System.Logger.Level.ERROR, "Failed to sync theme", (Throwable) e);
-            }
-        }
-
-        JComponent getComponent() {
-            var appearanceModeButton = new JButton();
-            {
-                var iconSize = 10;
-                var toLightMode = darkMode_sun.of(iconSize, iconSize);
-                var toDarkMode = darkMode_moon.of(iconSize, iconSize);
-                Runnable syncIconUpdater = () -> {
-                    switch (manager.getPreferredThemeStyle().getColorToneRule()) {
-                        case DARK:
-                            appearanceModeButton.putClientProperty(TO_APPEARANCE, TO_LIGHT_LAF);
-                            appearanceModeButton.setIcon(toLightMode);
-                            break;
-                        case LIGHT:
-                            appearanceModeButton.putClientProperty(TO_APPEARANCE, TO_DARK_LAF);
-                            appearanceModeButton.setIcon(toDarkMode);
-                            break;
-                    }
-                };
-                manager.addListener(e -> syncIconUpdater.run());
-                appearanceModeButton.addPropertyChangeListener("enabled", e -> syncIconUpdater.run());
-                syncIconUpdater.run();
-
-                appearanceModeButton.addActionListener(e -> {
-                    toggleSync(false);
-                    switch (appearanceModeButton.getClientProperty(TO_APPEARANCE).toString()) {
-                        case TO_DARK_LAF:
-                            FlatDarculaLaf.setup();
-                            Colors.setDarkMode(true);
-                            appearanceModeButton.putClientProperty(TO_APPEARANCE, TO_LIGHT_LAF);
-                            appearanceModeButton.setIcon(toLightMode);
-                            break;
-                        case TO_LIGHT_LAF:
-                            FlatIntelliJLaf.setup();
-                            Colors.setDarkMode(false);
-                            appearanceModeButton.putClientProperty(TO_APPEARANCE, TO_DARK_LAF);
-                            appearanceModeButton.setIcon(toDarkMode);
-                            break;
-                    }
-                    FlatLaf.updateUI();
-                    FlatAnimatedLafChange.hideSnapshotWithAnimation();
-                });
-            }
-            appearanceModeButton.setEnabled(false);
-
-            var syncAppearanceButton = new JCheckBox("Sync appearance");
-            {
-                syncAppearanceButton.addActionListener(e -> {
-                    toggleSync(syncAppearanceButton.isSelected());
-                    appearanceModeButton.setEnabled(!syncAppearanceButton.isSelected());
-                    SYNC_THEME_CHANGER.run();
-                });
-                syncAppearanceButton.setSelected(true);
-            }
-
-            var appearanceControlsPanel = new JPanel(new FlowLayout());
-            appearanceControlsPanel.add(appearanceModeButton);
-            appearanceControlsPanel.add(syncAppearanceButton);
-
-            return appearanceControlsPanel;
-        }
-    }
-
-    private static void installAppIcon(JFrame jFrame) {
-        var resource = FirePlaceMain.class.getClassLoader().getResource("fire.png");
-        var image = Toolkit.getDefaultToolkit().getImage(resource);
-
-        try {
-            Taskbar.getTaskbar().setIconImage(image);
-        } catch (UnsupportedOperationException | SecurityException ignored) {
-        }
-        jFrame.setIconImage(image);
-    }
 }
